% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readaqs.R
\name{read.aqs}
\alias{read.aqs}
\title{Read in and format raw data transactions from Air Quality Systems}
\usage{
read.aqs(file, level = 2, time.zone = "UTC", remove = FALSE,
  RC = "ignore", ...)
}
\arguments{
\item{file}{A character string of the name of the raw data file. If the
file is not in the working directory, the file path must be
specified in this argument (see example below).}

\item{level}{An integer between 0 and 4, specifying the level of simplification
          and formatting to be applied. Default is level 2.
          \itemize{
             \item \strong{Level 0} simply returns the file as a data frame.
             \item \strong{Level 1} converts the Date and Start.Time columns
                   to a single Date.Time column of class POSIXct. Drops: Date,
                   Start.Time, RD, Action.Code, Null.Data.Code, Sampling.Frequency,
                   Monitor.Protocol..MP..ID, Alternate.Method.Detectable.Limit,
                   Uncertainty, and Qualifier columns.
             \item \strong{Level 2} concatenates all monitor ID columns (State.Code,
                   County.Code, Site.ID, Parameter, POC, Sample.Duration, Unit,
                   Method), separated by a "-". Returns only Date.Time, concatenated
                   monitor IDs and Sample.Value columns.
             \item \strong{Level 3} applies names to each monitor ID. Example:
                   monitor ID "06-079-0005-44201-1-1-8-87" becomes
                   "CA-San_Luis_Obispo-0005-ozone-1-1_hour-ppb-model_400_ozone_analyzer".
                   Monitor names used in this function are available
                   \href{https://www.epa.gov/aqs/aqs-code-list}{here} on the AQS site.
                   This level returns only Date.Time, named monitor IDs and
                   Sample.Value columns.
             \item \strong{Level 4} converts level 3 data to wide format with
                   Date.Time and named monitor IDs as the columns, and Sample.Value
                   as the values in each row.
}}

\item{time.zone}{A character string specifying the time zone to be used when
date values are converted to POSIXct class in level 1. Default
is set to UTC. See as.POSIXct documentation for more time zone options.}

\item{remove}{A logical value indicating if columns which contain only a
single value should be dropped. Default is set to FALSE. If TRUE, this
option will be applied at all levels.}

\item{RC}{A character string indicating if rows corresponding to RC should be dropped or
returned. Can take either "ignore" (the default) or "include". If set to
ignore, all rows corresponding to RC will be dropped. If set to "include",
a list will be returned for each level containing a data frame for RD transactions
in the first index, and a data frame of RC transactions in the second index.}

\item{...}{Other parameters that will be passed to read.table.}
}
\value{
Returns either a single data frame or a list of two data frames.
}
\description{
Reads in a .txt file of raw data transactions with the option for specifying
different levels of formatting.
}
\examples{
aqs <- read.aqs(file = "Desktop/AMP501_1595753-0.txt",
                level = 4,
                remove = TRUE)


}
